Added Admin and Home controllers along with admin and home views. Also added layoutadmin.ejs with <!--SCRIPTS_ADMIN--> tags that will be used in the next commit step.
Added jquery js file and knockout js files that will be used in the next commit step.

The idea is for the js folder contents to be inserted between <!--SCRIPTS--> tags (sails does this by default) and the jsadmin folder contents to be inserted between the <!--SCRIPTS_ADMIN--> tags (will be done in the next couple of commits.